Initialize Data model with URL
Since the purpose of the YOCaml website is also to demonstrate the
ergonomics and versatility of YOCaml, I am recoding the models (which
should probably be bundled as a dedicated library to enrich the
archetypes). The model is heavily inspired by Kane [1] (once again) but
relies much more on the Uri module (to avoid having to do heavy work
on validation and generation). URLs are generally easy to use, making
it simple to implement the concept of a repository and build easy
editing links, as on Capsule [2], once again.

[1] https://github.com/xvw/kane
[2] https://github.com/xvw/capsule

Co-authored-by: Xavier Van de Woestyne <xaviervdw@gmail.com>