Add model for Repository
Once again, we were greatly inspired by Kane's code. For the moment,
we only support Github, Gitlab, Tangled, and Codeberg, because that's
what the site needs (so Github would suffice), but in the near future,
we would like to extract the models in the form of a library and we
would probably like to support more forges (and generic forges based
on link schemas). However, I think that for the needs of the YOCaml
website, the current support is sufficient.

Co-authored-by: Xavier Van de Woestyne <xaviervdw@gmail.com>